backend/pyproject.toml
[project]
name = "kanban-studio-backend"
version = "0.1.0"
description = "FastAPI backend for Kanban Studio"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-jose[cryptography]>=3.3",
    "aiosqlite>=0.21",
    "openai>=1.82",
    "python-dotenv>=1.1",
    "bcrypt>=4.0",
    "slowapi>=0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "anyio[trio]>=4.0",
    "pytest-anyio>=0.0.2",
    "httpx>=0.28",
]

[tool.pytest.ini_options]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
